Fix notification API erros

+++ src/main/java/com/nexters/rezoom/converter/domain/ConvertService.java
@@ -3,12 +3,11 @@ package com.nexters.rezoom.converter.domain;
 import com.nexters.rezoom.coverletter.domain.Coverletter;
 import com.nexters.rezoom.coverletter.domain.CoverletterRepository;
 import com.nexters.rezoom.member.domain.Member;
+import com.nexters.rezoom.util.FileUtils;
 import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
 
 import java.io.File;
-import java.io.FileOutputStream;
-import java.io.IOException;
 
 /**
  * Created by momentjin@gmail.com on 2019-08-29
@@ -25,28 +24,23 @@ public class ConvertService {
 
     public void convertFromFileToCoverletter(Member member, MultipartFile[] files) {
         for (MultipartFile multipartFile : files) {
-            File file = this.convertFile(multipartFile);
-            String fileExtension = file.getName().split("\\.")[1];
-            CoverletterConverter converter = ConverterFactory.createConverterByExtension(fileExtension, file);
-            Coverletter coverletter = converter.convert();
-            coverletter.setMember(member);
 
-            coverletterRepository.save(coverletter);
-            file.delete();
-        }
-    }
+            File file = null;
 
-    private File convertFile(MultipartFile multipartFile) {
-        try {
-            File convertFile = new File(multipartFile.getOriginalFilename());
-            convertFile.createNewFile();
+            try {
+                file = FileUtils.convertFile(multipartFile);
+                String fileExtension = FileUtils.getFileExtension(file);
 
-            try (FileOutputStream fos = new FileOutputStream(convertFile)) {
-                fos.write(multipartFile.getBytes());
+                CoverletterConverter converter = ConverterFactory.createConverterByExtension(fileExtension, file);
+                Coverletter coverletter = converter.convert();
+                coverletter.setMember(member);
+
+                coverletterRepository.save(coverletter);
+            } finally {
+                if (file != null) file.delete();
             }
-            return convertFile;
-        } catch (IOException e) {
-            throw new RuntimeException(e);
         }
     }
+
+
 }
